Increases checks when deploying DeploymentConfig
This PR:

- Adds more checks for IsDCRolledOut, including waiting for the new
revision, making sure that DeploymentConfig is actually deployed / has a
new image as well as the desired revision is available.
- More debug information when waiting for deploymentconfig
- Moves the functions from template.go to utils.go (I have no idea why a
utility function was in a templates file...)
- These changes also makes sure that our revision doesn't skip a number
/ goes to the next revision.

Closes https://github.com/openshift/odo/issues/1567
